requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.3.2
cryptography>=42.0.5
httpx>=0.27.0
pydantic>=2.6.4
python-dotenv>=1.0.1
